pyproject.toml
[project]
name = "datadoc"
version = "0.1.0"
description = "A modern command-line interface example"
authors = [
    {name = "Tony Qiu", email = "tongqing.qiu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "typer>=0.9.0,<0.10.0",
    "rich>=13.7.0,<14.0.0",  # Typer's recommended for rich output
    "pydantic>=2.6.0,<3.0.0",  # For data validation and settings management
    "pyyaml>=6.0.1,<7.0.0"  # For YAML file parsing
]

[project.scripts]
datadoc = "datadoc.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    {include = "datadoc"}
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
mypy = "^1.9.0"
pre-commit = "^3.6.0"
watchfiles = "^0.21.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
datamodel-code-generator = "^0.25.0"  # For generating Pydantic models from JSON schema
types-pyyaml = "^6.0.12.20250402"

[tool.ruff]
line-length = 120
target-version = "py311"
per-file-ignores = { "datadoc/models/odcs.py" = ["E501"] }

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.poetry.scripts]
datadoc = "datadoc.cli:main"
format = "poetry run ruff format ."
lint = "poetry run ruff check ."
lint-fix = "poetry run ruff check --fix ."
typecheck = "poetry run mypy ."

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=datadoc --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["datadoc"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "raise ImportError",
]
